moved generic neural net training & prediction code to nn_base.py to prepare for RNNs; don't use batch indexes as input parameters for neural network training, closes #42

+++ senti/transforms.py
@@ -139,4 +139,6 @@ class Clip(BaseEstimator):
         return self
 
     def transform(self, docs):
-        return vstack(np.hstack([doc[:self.max_len], np.zeros(max(0, self.max_len - len(doc)))]) for doc in docs)
+        return vstack(np.hstack([
+                doc[:self.max_len], np.zeros(max(0, self.max_len - len(doc)), dtype=doc.dtype)
+        ]) for doc in docs)
